requirements.txt
python-telegram-bot>=21.3
python-dotenv
google-genai
Pillow
Flask
